Renamed discovery package to internal in the SilvercrestWifiSocket binding. See openhab/static-code-analysis#100
