+++ addons/binding/org.openhab.binding.silvercrestwifisocket/OSGI-INF/SilvercrestWifiSocketDiscovery.xml
@@ -10,7 +10,7 @@
 
 -->
 <scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" immediate="true" modified="modified" name="org.openhab.binding.silvercrestwifisocket.discovery.SilvercrestWifiSocketDiscoveryService">
-   <implementation class="org.openhab.binding.silvercrestwifisocket.discovery.SilvercrestWifiSocketDiscoveryService"/>
+   <implementation class="org.openhab.binding.silvercrestwifisocket.internal.discovery.SilvercrestWifiSocketDiscoveryService"/>
    <service>
       <provide interface="org.eclipse.smarthome.config.discovery.DiscoveryService"/>
    </service>
