+++ addons/binding/org.openhab.binding.silvercrestwifisocket/src/main/java/org/openhab/binding/silvercrestwifisocket/internal/discovery/SilvercrestWifiSocketDiscoveryService.java
@@ -6,7 +6,7 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.openhab.binding.silvercrestwifisocket.discovery;
+package org.openhab.binding.silvercrestwifisocket.internal.discovery;
 
 import java.util.HashMap;
 import java.util.Map;
